Fix deltaTime implementation with proper integration
Refactored all update and physics methods to use float-based time deltas instead of nanoseconds. Introduced a Timer class for frame timing, capping delta to 0.1s. Replaced simSpeed with timeScale throughout the codebase. Updated Ball and Plane constructors and logic for new timing, including angular velocity for Plane. Adjusted debug visuals and initial velocities to match new units. Updated project files to include Timer. Cleaned up includes and ensured proper collider type initialization. UserInterface now displays and controls timeScale.

+++ Ball.h
@@ -8,7 +8,7 @@ class Ball : public Entity
 public:
 	Ball(float x, float y, float velocityX, float velocityY, float mass, float radius, Color color);
 	void draw() override;
-	void update(long long dt) override;
+	void update(float dt) override;
 	void handleWallCollisions() override;
 	float getRadius();
 private:
